+++ src/icalendar/tests/fuzzed/__init__.py
@@ -5,21 +5,54 @@ These test cases reproduce the failure.
 Some more tests can be added to make sure that the behavior works properly.
 """
 
+_value_error_matches = [
+    "component",
+    "parse",
+    "Expected",
+    "Wrong date format",
+    "END encountered",
+    "vDDD",
+    "recurrence",
+    "Offset must",
+    "Invalid iCalendar",
+    "alue MUST",
+    "Key name",
+    "Invalid content line",
+    "does not exist",
+    "base 64",
+    "must use datetime",
+    "Unknown date type",
+    "Wrong",
+    "Start time",
+    "iCalendar",
+    "recurrence",
+    "float, float",
+    "utc offset",
+    "parent",
+    "MUST be a datetime",
+    "Invalid month:",
+]
 
-def fuzz_calendar_v1(
+
+def fuzz_v1_calendar(
     from_ical, calendar_string: str, multiple: bool, should_walk: bool
 ):
     """Take a from_ical function and reproduce the error.
 
     The calendar_string is a fuzzed input.
     """
-    cal = from_ical(calendar_string, multiple=multiple)
+    try:
+        cal = from_ical(calendar_string, multiple=multiple)
 
-    if not multiple:
-        cal = [cal]
-    for c in cal:
-        if should_walk:
-            for event in c.walk("VEVENT"):
-                event.to_ical()
-        else:
-            c.to_ical()
+        if not multiple:
+            cal = [cal]
+        for c in cal:
+            if should_walk:
+                for event in c.walk("VEVENT"):
+                    event.to_ical()
+            else:
+                c.to_ical()
+    except (ValueError, TypeError) as e:
+        if any(m in str(e) for m in _value_error_matches):
+            return -1
+        raise