+++ CHANGES.rst
@@ -55,6 +55,28 @@ New features
 - Added documentation of Vale usage for icalender documentation style guide. See `Issue 815 <https://github.com/collective/icalendar/issues/815>`_.
 - Added funding link to :file:`pyproject.toml`.  See `Issue 800 <https://github.com/collective/icalendar/issues/800>`_.
 - Added documentation of docstring format. See `Issue 747 <https://github.com/collective/icalendar/issues/747>`_.
+- Add ``TypeAlias`` ``icalendar.VPROPERTY`` as a type for all properties.
+- Allow access to :class:`icalendar.TypesFactory` via :func:`icalendar.TypesFactory.instance`.
+- Classes in :mod:`icalendar.prop` can all be initialized with parameters now.
+- Allow passing :class:`icalendar.vDDDTypes` as arguments to :class:`icalendar.vDDDLists`.
+- Add support for :class:`datetime.timezone` identification.
+- Add :func:`icalendar.is_utc` to check if a timezone is UTC.
+- Add support for JCal parsing and serialization. See `Issue 978 <https://github.com/collective/icalendar/issues/978>`_.
+
+  - Add :class:`icalendar.JCalParsingError` exception for jCal parsing.
+  - Add :attr:`icalendar.Parameters.tzid` and :func:`icalendar.Parameters.is_utc` for timezone handling.
+  - Add :class:`icalendar.vUnknown` which is the default now for unknown properties and the same as :class:`icalendar.vText`.
+  - Add :func:`icalendar.ComponentFactory.get_component_class` that also creates a new component class when it does not exist.
+  - Add :func:`icalendar.ComponentFactory.add_component_class` to have a unified interface to add component classes.
+  - Add :func:`icalendar.Component.to_json`, :func:`icalendar.Component.to_jcal`, and :func:`icalendar.Component.from_jcal` methods so all components can be serialized and parsed.
+  - Classes in :mod:`icalendar.prop` receive the following updates:
+
+    - Add ``to_jcal`` and ``from_jcal`` for serialization.
+    - Add ``default_value`` class attribute for the default ``VALUE`` parameter.
+    - Add ``VALUE`` parameter property.
+    - Add ``examples()`` classmethod that returns examples for testing.
+
+- Added documentation of how to fuzz icalendar. See `Issue 905 <https://github.com/collective/icalendar/issues/905>`_.
 - Added first-time contributors section to Contributing. See `Issue 934 <https://github.com/collective/icalendar/issues/934>`_.
 - Revised check of which files trigger a documentation build on Read the Docs. See `Issue 848 <https://github.com/collective/icalendar/issues/848>`_.
 - Created a separate path to contribute to documentation. See `Issue 929 <https://github.com/collective/icalendar/issues/929>`_.
@@ -66,6 +88,8 @@ New features
 Bug fixes
 ~~~~~~~~~
 
+- Correct ``name`` of ``AVAILABLE`` component.
+- Fix ical serialization of :class:`icalendar.vTime`.
 - Make sure a component's ``conferences`` property adds ``VALUE=URI`` if the conference is created as a URI.
 - Fix CLI deprecated :py:class:`argparse.FileType` usage and remove ``PendingDeprecationWarning``. See `Issue #1005 <https://github.com/collective/icalendar/issues/1005>`_.